categories/serializers.py
from rest_framework.serializers import ModelSerializer
from . import models


class CategorySerializer(ModelSerializer):
    
    class Meta:
        model = models.Category
        fields = '__all__'